ABACProjectDomain1/PAP/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-xml==2.0.0
djongo==1.3.3
idna==2.10
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
sqlparse==0.2.4
urllib3==1.25.10
xmltodict==0.12.0
